requirements.txt
# ============================================================
# Financial Document Analyser — requirements.txt
#
# STRATEGY: Only pin TOP-LEVEL packages we directly use.
# Let pip resolve all transitive dependencies automatically.
# ============================================================

# --- Core framework ---
crewai==0.130.0
crewai-tools==0.38.0

# --- LLM providers ---
groq>=0.9.0

# --- API server ---
fastapi>=0.110.0
# BUG FIX: uvicorn was missing — needed to serve the FastAPI app
uvicorn>=0.29.0
python-multipart>=0.0.9

# --- PDF reading ---
# BUG FIX: pypdf was missing — needed by tools.py to read PDFs
pypdf>=4.2.0

# --- Utilities ---
python-dotenv>=1.0.0

# --- numpy pin (avoids binary incompatibility on some systems) ---
numpy>=1.26.0,<2.0.0

# ── BONUS: Queue Worker ──────────────────────────────────────
# Celery: distributed task queue
celery>=5.4.0
# Redis client: broker + result backend for Celery
redis>=5.0.7

# ── BONUS: Database Integration ─────────────────────────────
# SQLAlchemy ORM (works with SQLite for dev, PostgreSQL for prod)
sqlalchemy>=2.0.0
# PostgreSQL driver (only needed if DATABASE_URL is postgres://)
# psycopg2-binary>=2.9.9